@@ -12,27 +12,29 @@
   "author": "Sempris",
   "license": "UNLICENSED",
   "dependencies": {
-    "@material-ui/core": "4.11.0",
-    "@material-ui/icons": "4.9.1",
-    "@material-ui/styles": "4.10.0",
+    "@emotion/react": "11.5.0",
+    "@emotion/styled": "11.3.0",
+    "@mui/icons-material": "5.0.4",
+    "@mui/material": "5.0.4",
+    "@mui/styles": "^5.0.1",
     "aws-amplify": "3.3.3",
-    "axios": "0.20.0",
-    "i18next": "19.8.2",
+    "axios": "0.23.0",
+    "i18next": "21.3.3",
     "i18next-browser-languagedetector": "6.0.1",
     "i18next-xhr-backend": "3.2.2",
-    "immutable": "4.0.0-rc.12",
-    "notistack": "1.0.1",
-    "password-validator": "^5.1.1",
+    "immutable": "4.0.0",
+    "notistack": "2.0.2",
+    "password-validator": "5.1.1",
     "prop-types": "15.7.2",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
     "react-grid-layout": "1.1.1",
-    "react-i18next": "11.7.3",
-    "react-player": "^2.9.0",
-    "react-redux": "7.2.1",
-    "react-router-dom": "5.2.0",
-    "recharts": "2.0.0-beta.7",
-    "redux": "4.0.5",
+    "react-i18next": "11.12.0",
+    "react-player": "2.9.0",
+    "react-redux": "7.2.5",
+    "react-router-dom": "5.3.0",
+    "recharts": "2.1.5",
+    "redux": "4.1.1",
     "redux-actions": "2.6.5",
     "redux-observable": "1.2.0",
     "rxjs": "6.6.3",
@@ -40,7 +42,13 @@
   },
   "devDependencies": {
     "prettier": "2.1.2",
-    "react-scripts": "3.4.3"
+    "react-scripts": "4.0.3",
+    "eslint": "7.11.0",
+    "eslint-config-standard": "16.0.3",
+    "eslint-plugin-import": "2.25.2",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "5.1.1",
+    "eslint-plugin-react": "7.26.1"
   },
   "browserslist": {
     "production": [
