@@ -1,17 +1,23 @@
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
-import { ThemeProvider } from '@material-ui/core';
-import CssBaseline from '@material-ui/core/CssBaseline';
+import { ThemeProvider as MuiThemeProvider } from '@mui/material/styles';
+import { StyledEngineProvider } from '@mui/styled-engine';
+import { ThemeProvider } from '@emotion/react';
+import CssBaseline from '@mui/material/CssBaseline';
 import defaultTheme from './defaultTheme';
 
 const Theme = ({ children }) => {
 	return (
-		<ThemeProvider theme={defaultTheme}>
-			<Fragment>
-				<CssBaseline />
-				{children}
-			</Fragment>
-		</ThemeProvider>
+		<StyledEngineProvider injectFirst>
+			<MuiThemeProvider theme={defaultTheme}>
+				<ThemeProvider theme={defaultTheme}>
+					<Fragment>
+						<CssBaseline />
+						{children}
+					</Fragment>
+				</ThemeProvider>
+			</MuiThemeProvider>
+		</StyledEngineProvider>
 	);
 };
 
