@@ -4,8 +4,8 @@ import PropTypes from 'prop-types';
 import { Auth } from 'aws-amplify';
 import { withTranslation } from 'react-i18next';
 
-import { withStyles } from '@material-ui/styles';
-import Button from '@material-ui/core/Button';
+import { withStyles } from '@mui/styles';
+import Button from '@mui/material/Button';
 import { userActions } from '../../containers/User';
 
 const styles = (theme) => ({
